etl/requirements.txt
sec-edgar-downloader==5.0.3
beautifulsoup4==4.12.3
psycopg2-binary==2.9.9
vaderSentiment==3.3.2
requests==2.32.2
python-dotenv==1.0.1
pandas==2.2.2
